Задание 5.5.5 - тип данных Decimal как самый большой

+++ Module_05/Quest_555/Quest_555.cs
@@ -8,7 +8,7 @@ namespace Module_05.Quest_555
 {
     internal class Quest_555
     {
-        static int Factorial(int x)
+        static decimal Factorial(int x)
         {
             if (x == 0)
             {
